requirements.txt
pyvi==0.1.1
pinecone-client
langchain==1.2.12
langchain-community==0.4.1
langchain-pinecone==0.2.13
langchain-tavily==0.2.17
python-dotenv
jupyter==1.1.1
pandas==2.3.3
seaborn==0.13.2
huggingface_hub==1.8.0
langgraph==1.1.3
transformers==5.3.0
torch==2.11.0
pydantic==2.12.5
fastapi==0.135.2
uvicorn==0.42.0
pypdf==6.9.2
ollama==0.6.1
fastapi==0.135.2
langchain-ollama==0.3.0